Rename nnmaildir-request-marks back to nnmaildir-request-update-info
This function was renamed as part of a greater change on 2010-09-23,
but nnmaildir actually needs this function to be called to be able to
read marks from the maildir (as opposed to from .newsrc.eld).  As
noted in the file comments, a goal of nnmaildir is to have all
information about a group stored in the maildir.

* nnmaildir.el (nnmaildir-request-update-info): Rename from
nnmaildir-request-marks.