Make nnmaildir understand and write maildir flags
That is, rename files from "unique:2," to "unique:2,S" for "seen", etc.
This should make nnmaildir more usable with offlineimap.

* nnmaildir.el (nnmaildir-flag-mark-mapping): New constant.
(nnmaildir--mark-to-flag, nnmaildir--flag-to-mark)
(nnmaildir--ensure-suffix, nnmaildir--add-flag)
(nnmaildir--remove-flag, nnmaildir--article-set-flags): New functions.
(nnmaildir--scan): Don't blindly append ":2,"; call
`nnmaildir--ensure-suffix' to ensure idempotency.
When counting unseen and ticked articles, consider flags
in file names.
(nnmaildir-request-update-info): Look for flags in file names as well
as in mark directories.
(nnmaildir-request-set-mark): When clearing a mark, remove the
corresponding flag as well, if any.  When setting a mark corresponding
to a flag, just rename, don't use the mark directory.